// packages/ollama-connector/src/types.d.ts
export interface OllamaMessage {
    role: string;
    content: string;
    thinking?: string;
    images?: Uint8Array[] | string[];
    tool_calls?: any[];
    tool_name?: string;
}